WIP. Add reading files to BufferedImage

+++ src/main/java/com/justalex/pnp/HttpWorker.groovy
@@ -7,18 +7,23 @@ import java.nio.file.Paths
 
 class HttpWorker {
 
-    void downloadImagesForCards(List cards, Path pathToFolderToSaveIn, CardType cardType) {
-        println("Downloading [$cards.size][$cardType.name] cards")
+    void downloadImagesForCards(List cards, Path pathToFolderToSaveIn, String cardType) {
+        println("Downloading [$cards.size][$cardType] cards")
+        def imageWorker = new ImageWorker()
         cards.indexed().each { index, card ->
             def card_id = card.id
             def warbandId = card.warbandId
             println("$index: $card_id")
-            Paths.get().
-            def pathToFile = Paths.get("card_images/power_cards/cards/$warbandId/${card_id}.png")
+            def pathToFile = pathToFolderToSaveIn.resolve("${card_id}.png")
+//            def pathToFile = Paths.get("card_images/power_cards/cards/$warbandId/${card_id}.png")
+
             Files.createDirectories(pathToFile.getParent())
-            Files.createFile(pathToFile)
-            Files.write(pathToFile, imageWorker.getBytesFromUrl(card.image_url))
+            if (Files.notExists(pathToFile)) {
+                Files.createFile(pathToFile)
+                Files.write(pathToFile, imageWorker.getBytesFromUrl(card.image_url))
+            }
         }
+        println("Done")
     }
 
 }
